data_acquisition/backend/requirements.txt
uvicorn==0.17.5
kopf==1.35.3
uvloop==0.16.0
fastapi==0.75.0
pykube-ng==22.1.0
SQLAlchemy==1.4.35
## The following requirements were added by pip freeze:
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.0.4
frozenlist==1.3.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
multidict==6.0.2
pydantic==1.9.0
python-json-logger==2.0.2
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
yarl==1.7.2
certifi==2021.10.8
requests==2.27.1
urllib3==1.26.9